+++ packages/analytics-compat/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @firebase/analytics-compat
 
+## 0.1.12
+
+### Patch Changes
+
+- Updated dependencies [[`835f1d46a`](https://github.com/firebase/firebase-js-sdk/commit/835f1d46a6780535bc660ef7dc23293350d5fe43), [`efe2000fc`](https://github.com/firebase/firebase-js-sdk/commit/efe2000fc499e2c85c4e5e0fef6741ff3bad2eb0)]:
+  - @firebase/analytics@0.7.11
+  - @firebase/util@1.6.2
+  - @firebase/component@0.5.16
+
 ## 0.1.11
 
 ### Patch Changes