+++ packages/app-check-compat/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @firebase/app-check-compat
 
+## 0.2.10
+
+### Patch Changes
+
+- Updated dependencies [[`efe2000fc`](https://github.com/firebase/firebase-js-sdk/commit/efe2000fc499e2c85c4e5e0fef6741ff3bad2eb0)]:
+  - @firebase/app-check@0.5.10
+  - @firebase/util@1.6.2
+  - @firebase/component@0.5.16
+
 ## 0.2.9
 
 ### Patch Changes